Feat: redirect to the main page when not logged in on my page, reservation page, reservation completed page
